Move `/react/components/Icons` to `/react/Icons`
